Only print to-the-second resolution by default.
Put microseconds behind a flag.

This makes it harder to do timing analysis on the log file to try to
turn this into a keylogger.

(More precise timing information is still available to an on-line attack
by watching the flush timings or by watching the CPU usage and run state
of the process.)